docs: create module for bitvavo api
